Get Predictions for Statistical Models

A generic function to get predictions for statistical models.

Usage

get_predictions(model.obj, classify, pred.obj = NULL, vars = NULL, ...)

# S3 method for class 'asreml'
get_predictions(model.obj, classify, pred.obj = NULL, vars = NULL, ...)

# S3 method for class 'lm'
get_predictions(model.obj, classify, ...)

# S3 method for class 'lmerMod'
get_predictions(model.obj, classify, ...)

# S3 method for class 'lmerModLmerTest'
get_predictions(model.obj, classify, ...)

Arguments

model.obj

A mixed-effects model object (lmerMod or lmerModLmerTest).

classify

Name of predictor variable as a string.

pred.obj

Optional precomputed prediction object.

vars

Variables used in the model.

...

Additional arguments passed to emmeans::emmeans.

Value

A list containing predictions, standard errors, degrees of freedom, response variable label, and aliased names.

A list containing predictions, standard errors, degrees of freedom, response variable label, and aliased names.

A list containing predictions, standard errors, degrees of freedom, response variable label, and aliased names.

A list containing predictions, standard errors, degrees of freedom, response variable label, and aliased names.